+++ NEWS.md
@@ -2,6 +2,7 @@
 * Added option for an explicit Missing Category with custom cuts
 * Added option to compute the inverse of Numeric variables
 * Moved the rounding to after division/inverse
+* Added possibility to round data labels before plotting
 
 # ggquickeda 0.1.7
 * Added support for semicolon separated files and added package version in App Title